Consolidate the many copies of NewTestCAWithConfig()
