+++ lib/auth/authtest/ca.go
@@ -1,0 +1,38 @@
+// Teleport
+// Copyright (C) 2025 Gravitational, Inc.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+package authtest
+
+import (
+	"github.com/gravitational/teleport/api/types"
+	"github.com/gravitational/teleport/lib/auth/authcatest"
+)
+
+// TestCAConfig defines the configuration for generating
+// a test certificate authority
+type TestCAConfig = authcatest.TestCAConfig
+
+// NewTestCA returns new test authority with a test key as a public and
+// signing key
+func NewTestCA(caType types.CertAuthType, clusterName string, privateKeys ...[]byte) *types.CertAuthorityV2 {
+	return authcatest.NewTestCA(caType, clusterName, privateKeys...)
+}
+
+// NewTestCAWithConfig generates a new certificate authority with the specified
+// configuration
+func NewTestCAWithConfig(config TestCAConfig) *types.CertAuthorityV2 {
+	return authcatest.NewTestCAWithConfig(config)
+}
